requirements.txt
asgiref==3.10.0
brotli==1.2.0
cffi==2.0.0
charset-normalizer==3.4.4
cssselect2==0.8.0
Django==5.2.8
django-simple-history==3.11.0
fonttools==4.61.1
pillow==12.1.0
pycparser==2.23
pydyf==0.12.1
pyphen==0.17.2
reportlab==4.4.7
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.2
webencodings==0.5.1
zopfli==0.4.0